Split up Damageable test into smaller file and set fields to private
https://github.com/leobenkel/PSF-LoginServer/issues/2
https://github.com/leobenkel/PSF-LoginServer/issues/1
